+++ app/controllers/application_controller.rb
@@ -6,8 +6,4 @@ class ApplicationController < ActionController::Base
 
   responders :flash
   respond_to :html
-
-  decent_configuration do
-    strategy DecentExposure::StrongParametersStrategy
-  end
 end