+++ lib/templates/rails/scaffold_controller/controller.rb
@@ -2,16 +2,16 @@
 class <%= controller_class_name %>Controller < ApplicationController
   respond_to :html
 
-  expose(:<%= singular_table_name %>, attributes: :<%= singular_table_name %>_params)
-  expose(:<%= plural_table_name %>) { <%= class_name %>.page(params[:page]) }
+  expose :<%= singular_table_name %>
+  expose :<%= plural_table_name %>, -> { <%= class_name %>.page(params[:page]) }
 
   def create
-    flash[:notice] = '<%= human_name %> was successfully created.' if <%= singular_table_name %>.save
+    <%= singular_table_name %>.save
     respond_with(<%= singular_table_name %>)
   end
 
   def update
-    flash[:notice] = '<%= human_name %> was successfully updated.' if <%= singular_table_name %>.save
+    <%= singular_table_name %>.update_attributes(<%= singular_table_name %>_params)
     respond_with(<%= singular_table_name %>)
   end